adjustment to address editing
removed the route for editing an address and made editing in the same container
